Time: 15 ms (17.15%), Space: 23.1 MB (5.17%) - LeetHub

+++ 0875-koko-eating-bananas/0875-koko-eating-bananas.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     bool func(int mid,vector<int>& piles, int h){
         int n=piles.size();
-        int hours=0;
+        long long hours=0;
         for(int i=0;i<n;i++){
             hours+=ceil((double)piles[i]/mid);
         }
